Move loadProto and saveProto into FSCommon
