+++ src/mesh/NodeDB.cpp
@@ -15,7 +15,6 @@
 #include "RTC.h"
 #include "Router.h"
 #include "SPILock.h"
-#include "SafeFile.h"
 #include "TypeConversions.h"
 #include "error.h"
 #include "main.h"
@@ -1105,39 +1104,6 @@ void NodeDB::pickNewNodeNum()
     myNodeInfo.my_node_num = nodeNum;
 }
 
-/** Load a protobuf from a file, return LoadFileResult */
-LoadFileResult NodeDB::loadProto(const char *filename, size_t protoSize, size_t objSize, const pb_msgdesc_t *fields,
-                                 void *dest_struct)
-{
-    LoadFileResult state = LoadFileResult::OTHER_FAILURE;
-#ifdef FSCom
-    concurrency::LockGuard g(spiLock);
-
-    auto f = FSCom.open(filename, FILE_O_READ);
-
-    if (f) {
-        LOG_INFO("Load %s", filename);
-        pb_istream_t stream = {&readcb, &f, protoSize};
-        if (fields != &meshtastic_NodeDatabase_msg) // contains a vector object
-            memset(dest_struct, 0, objSize);
-        if (!pb_decode(&stream, fields, dest_struct)) {
-            LOG_ERROR("Error: can't decode protobuf %s", PB_GET_ERROR(&stream));
-            state = LoadFileResult::DECODE_FAILED;
-        } else {
-            LOG_INFO("Loaded %s successfully", filename);
-            state = LoadFileResult::LOAD_SUCCESS;
-        }
-        f.close();
-    } else {
-        LOG_ERROR("Could not open / read %s", filename);
-    }
-#else
-    LOG_ERROR("ERROR: Filesystem not implemented");
-    state = LoadFileResult::NO_FILESYSTEM;
-#endif
-    return state;
-}
-
 void NodeDB::loadFromDisk()
 {
     // Mark the current device state as completely unusable, so that if we fail reading the entire file from
@@ -1349,34 +1315,6 @@ void NodeDB::loadFromDisk()
     }
 
 #endif
-}
-
-/** Save a protobuf from a file, return true for success */
-bool NodeDB::saveProto(const char *filename, size_t protoSize, const pb_msgdesc_t *fields, const void *dest_struct,
-                       bool fullAtomic)
-{
-    bool okay = false;
-#ifdef FSCom
-    auto f = SafeFile(filename, fullAtomic);
-
-    LOG_INFO("Save %s", filename);
-    pb_ostream_t stream = {&writecb, static_cast<Print *>(&f), protoSize};
-
-    if (!pb_encode(&stream, fields, dest_struct)) {
-        LOG_ERROR("Error: can't encode protobuf %s", PB_GET_ERROR(&stream));
-    } else {
-        okay = true;
-    }
-
-    bool writeSucceeded = f.close();
-
-    if (!okay || !writeSucceeded) {
-        LOG_ERROR("Can't write prefs!");
-    }
-#else
-    LOG_ERROR("ERROR: Filesystem not implemented");
-#endif
-    return okay;
 }
 
 bool NodeDB::saveChannelsToDisk()
